fix circular reference and increase icon size
