+++ src/pageEditor/tabs/editTab/editTabTypes.ts
@@ -23,9 +23,12 @@
  * Keep in mind that despite it looks like an array (the top-level may look like an array - have numbers for property names), it is an object.
  * For instance, it doesn't have a `length` property.
  */
-import { BrickNodeContentProps } from "@/pageEditor/tabs/editTab/editorNodes/brickNode/BrickNodeContent";
+
 import { MoveBrickControlProps } from "@/pageEditor/tabs/editTab/editorNodes/brickNode/MoveBrickControl";
 import { NodeAction } from "@/pageEditor/tabs/editTab/editorNodes/nodeActions/NodeActionsView";
+import { IconProp } from "@fortawesome/fontawesome-svg-core";
+import React from "react";
+import { OutputKey } from "@/core";
 
 export type FormikError = string | FormikErrorTree;
 
@@ -41,6 +44,13 @@ export enum RunStatus {
   WARNING,
   ERROR,
 }
+
+export type BrickNodeContentProps = {
+  icon?: IconProp | React.ReactNode;
+  runStatus?: RunStatus;
+  brickLabel: string;
+  outputKey?: OutputKey;
+};
 
 export type BrickNodeProps = BrickNodeContentProps &
   MoveBrickControlProps & {
