+++ src/pageEditor/tabs/editTab/editorNodes/brickNode/BrickNodeContent.module.scss
@@ -32,6 +32,8 @@
   width: 1.1em;
   margin-right: 10px;
 
+  font-size: 1rem;
+
   // External icon
   img {
     // !important is to override the inline CSS
